refactor ingredient game package structure and update imports
