+++ src/main/kotlin/di/UIModule.kt
@@ -1,6 +1,7 @@
 package org.example.di
 
 import org.example.presentation.*
+import org.example.presentation.game.IngredientGameUI
 import org.example.presentation.suggest.IraqiMealsUI
 import org.example.presentation.suggest.EggFreeSweetsUI
 import org.koin.dsl.bind
